Extract to commandlib/commandfinder.ts
