+++ src/providers/completer/command.ts
@@ -4,7 +4,8 @@ import {latexParser} from 'latex-utensils'
 
 import {Extension} from '../../main'
 import {Environment, EnvSnippetType} from './environment'
-import {IProvider} from './interface'
+import type {IProvider} from './interface'
+import {CommandFinder, isTriggerSuggestNeeded} from './commandlib/commandfinder'
 
 interface CmdItemEntry {
     command: string, // frame
@@ -20,18 +21,14 @@ export interface Suggestion extends vscode.CompletionItem {
     package: string
 }
 
-function isTriggerSuggestNeeded(name: string): boolean {
-    const reg = /[a-z]*(cite|ref|input)[a-z]*|begin|bibitem|(sub)?(import|includefrom|inputfrom)|gls(?:pl|text|first|plural|firstplural|name|symbol|desc|user(?:i|ii|iii|iv|v|vi))?|Acr(?:long|full|short)?(?:pl)?|ac[slf]?p?/i
-    return reg.test(name)
-}
 
 export class Command implements IProvider {
     private readonly extension: Extension
     private readonly environment: Environment
+    private readonly commandFinder: CommandFinder
 
     packages: string[] = []
     bracketCmds: {[key: string]: Suggestion} = {}
-    definedCmds: {[key: string]: {file: string, location: vscode.Location}} = {}
     private defaultCmds: Suggestion[] = []
     private defaultSymbols: Suggestion[] = []
     private packageCmds: {[pkg: string]: Suggestion[]} = {}
@@ -39,6 +36,7 @@ export class Command implements IProvider {
     constructor(extension: Extension, environment: Environment) {
         this.extension = extension
         this.environment = environment
+        this.commandFinder = new CommandFinder()
     }
 
     initialize(defaultCmds: {[key: string]: CmdItemEntry}) {
@@ -60,6 +58,10 @@ export class Command implements IProvider {
         this.environment.getDefaultEnvs(EnvSnippetType.AsCommand).forEach(cmd => {
             this.defaultCmds.push(cmd)
         })
+    }
+
+    get definedCmds() {
+        return this.commandFinder.definedCmds
     }
 
     provideFrom(_type: string, result: RegExpMatchArray, args: {document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext}) {
@@ -218,9 +220,9 @@ export class Command implements IProvider {
             }
         })
         if (nodes !== undefined) {
-            this.extension.manager.cachedContent[file].element.command = this.getCmdFromNodeArray(file, nodes)
+            this.extension.manager.cachedContent[file].element.command = this.commandFinder.getCmdFromNodeArray(file, nodes)
         } else if (content !== undefined) {
-            this.extension.manager.cachedContent[file].element.command = this.getCmdFromContent(file, content)
+            this.extension.manager.cachedContent[file].element.command = this.commandFinder.getCmdFromContent(file, content)
         }
     }
 
@@ -238,14 +240,6 @@ export class Command implements IProvider {
             return i > -1 ? name.substr(0, i): name
         }
         return name
-    }
-
-    private getCmdFromNodeArray(file: string, nodes: latexParser.Node[], cmdList: string[] = []): Suggestion[] {
-        let cmds: Suggestion[] = []
-        nodes.forEach(node => {
-            cmds = cmds.concat(this.getCmdFromNode(file, node, cmdList))
-        })
-        return cmds
     }
 
     getExtraPkgs(languageId: string): string[] {
@@ -336,192 +330,6 @@ export class Command implements IProvider {
         })
     }
 
-    private getCmdFromNode(file: string, node: latexParser.Node, cmdList: string[] = []): Suggestion[] {
-        const cmds: Suggestion[] = []
-        if (latexParser.isDefCommand(node)) {
-           const name = node.token.slice(1)
-            if (!cmdList.includes(name)) {
-                const cmd: Suggestion = {
-                    label: `\\${name}`,
-                    kind: vscode.CompletionItemKind.Function,
-                    documentation: '`' + name + '`',
-                    insertText: new vscode.SnippetString(name + this.getArgsFromNode(node)),
-                    filterText: name,
-                    package: ''
-                }
-                cmds.push(cmd)
-                cmdList.push(name)
-            }
-        } else if (latexParser.isCommand(node)) {
-            if (!cmdList.includes(node.name)) {
-                const cmd: Suggestion = {
-                    label: `\\${node.name}`,
-                    kind: vscode.CompletionItemKind.Function,
-                    documentation: '`' + node.name + '`',
-                    insertText: new vscode.SnippetString(node.name + this.getArgsFromNode(node)),
-                    filterText: node.name,
-                    package: ''
-                }
-                if (isTriggerSuggestNeeded(node.name)) {
-                    cmd.command = { title: 'Post-Action', command: 'editor.action.triggerSuggest' }
-                }
-                cmds.push(cmd)
-                cmdList.push(node.name)
-            }
-            if (['newcommand', 'renewcommand', 'providecommand', 'DeclarePairedDelimiter', 'DeclarePairedDelimiterX', 'DeclarePairedDelimiterXPP'].includes(node.name) &&
-                Array.isArray(node.args) && node.args.length > 0) {
-                const label = (node.args[0].content[0] as latexParser.Command).name
-                let args = ''
-                if (latexParser.isOptionalArg(node.args[1])) {
-                    const numArgs = parseInt((node.args[1].content[0] as latexParser.TextString).content)
-                    for (let i = 1; i <= numArgs; ++i) {
-                        args += '{${' + i + '}}'
-                    }
-                }
-                if (!cmdList.includes(label)) {
-                    const cmd: Suggestion = {
-                        label: `\\${label}`,
-                        kind: vscode.CompletionItemKind.Function,
-                        documentation: '`' + label + '`',
-                        insertText: new vscode.SnippetString(label + args),
-                        filterText: label,
-                        package: 'user-defined'
-                    }
-                    cmds.push(cmd)
-                    this.definedCmds[label] = {
-                        file,
-                        location: new vscode.Location(
-                            vscode.Uri.file(file),
-                            new vscode.Position(node.location.start.line - 1, node.location.start.column))
-                    }
-                    cmdList.push(label)
-                }
-            }
-        }
-        if (latexParser.hasContentArray(node)) {
-            return cmds.concat(this.getCmdFromNodeArray(file, node.content, cmdList))
-        }
-        return cmds
-    }
-
-    private getArgsFromNode(node: latexParser.Node): string {
-        let args = ''
-        if (!('args' in node)) {
-            return args
-        }
-        let index = 0
-        if (latexParser.isCommand(node)) {
-            node.args.forEach(arg => {
-                ++index
-                if (latexParser.isOptionalArg(arg)) {
-                    args += '[${' + index + '}]'
-                } else {
-                    args += '{${' + index + '}}'
-                }
-            })
-            return args
-        }
-        if (latexParser.isDefCommand(node)) {
-            node.args.forEach(arg => {
-                ++index
-                if (latexParser.isCommandParameter(arg)) {
-                    args += '{${' + index + '}}'
-                }
-            })
-            return args
-        }
-        return args
-    }
-
-    private getCmdFromContent(file: string, content: string): Suggestion[] {
-        const cmdReg = /\\([a-zA-Z@_]+(?::[a-zA-Z]*)?)({[^{}]*})?({[^{}]*})?({[^{}]*})?/g
-        const cmds: Suggestion[] = []
-        const cmdList: string[] = []
-        let explSyntaxOn: boolean = false
-        while (true) {
-            const result = cmdReg.exec(content)
-            if (result === null) {
-                break
-            }
-            if (result[1] === 'ExplSyntaxOn') {
-                explSyntaxOn = true
-                continue
-            } else if (result[1] === 'ExplSyntaxOff') {
-                explSyntaxOn = false
-                continue
-            }
-
-
-            if (!explSyntaxOn) {
-                const len = result[1].search(/[_:]/)
-                if (len > -1) {
-                    result[1] = result[1].slice(0, len)
-                }
-            }
-            if (cmdList.includes(result[1])) {
-                continue
-            }
-            const cmd: Suggestion = {
-                label: `\\${result[1]}`,
-                kind: vscode.CompletionItemKind.Function,
-                documentation: '`' + result[1] + '`',
-                insertText: new vscode.SnippetString(this.getArgsFromRegResult(result)),
-                filterText: result[1],
-                package: ''
-            }
-            if (isTriggerSuggestNeeded(result[1])) {
-                cmd.command = { title: 'Post-Action', command: 'editor.action.triggerSuggest' }
-            }
-            cmds.push(cmd)
-            cmdList.push(result[1])
-        }
-
-        const newCommandReg = /\\(?:(?:(?:re|provide)?(?:new)?command)|(?:DeclarePairedDelimiter(?:X|XPP)?))(?:{)?\\(\w+)/g
-        while (true) {
-            const result = newCommandReg.exec(content)
-            if (result === null) {
-                break
-            }
-            if (cmdList.includes(result[1])) {
-                continue
-            }
-
-            const cmd: Suggestion = {
-                label: `\\${result[1]}`,
-                kind: vscode.CompletionItemKind.Function,
-                documentation: '`' + result[1] + '`',
-                insertText: result[1],
-                filterText: result[1],
-                package: 'user-defined'
-            }
-            cmds.push(cmd)
-            cmdList.push(result[1])
-
-            this.definedCmds[result[1]] = {
-                file,
-                location: new vscode.Location(
-                    vscode.Uri.file(file),
-                    new vscode.Position(content.substr(0, result.index).split('\n').length - 1, 0))
-            }
-        }
-
-        return cmds
-    }
-
-    private getArgsFromRegResult(result: RegExpExecArray): string {
-        let text = result[1]
-
-        if (result[2]) {
-            text += '{${1}}'
-        }
-        if (result[3]) {
-            text += '{${2}}'
-        }
-        if (result[4]) {
-            text += '{${3}}'
-        }
-        return text
-    }
 
     private entryCmdToCompletion(itemKey: string, item: CmdItemEntry): Suggestion {
         const configuration = vscode.workspace.getConfiguration('latex-workshop')
@@ -602,4 +410,5 @@ export class Command implements IProvider {
             }
         })
     }
+
 }
