+++ src/providers/completer/commandlib/commandfinder.ts
@@ -1,0 +1,210 @@
+import * as vscode from 'vscode'
+import {latexParser} from 'latex-utensils'
+
+import type {Suggestion} from '../command'
+
+
+export function isTriggerSuggestNeeded(name: string): boolean {
+    const reg = /[a-z]*(cite|ref|input)[a-z]*|begin|bibitem|(sub)?(import|includefrom|inputfrom)|gls(?:pl|text|first|plural|firstplural|name|symbol|desc|user(?:i|ii|iii|iv|v|vi))?|Acr(?:long|full|short)?(?:pl)?|ac[slf]?p?/i
+    return reg.test(name)
+}
+
+export class CommandFinder {
+    definedCmds: {[key: string]: {file: string, location: vscode.Location}} = {}
+
+    getCmdFromNodeArray(file: string, nodes: latexParser.Node[], cmdList: string[] = []): Suggestion[] {
+        let cmds: Suggestion[] = []
+        nodes.forEach(node => {
+            cmds = cmds.concat(this.getCmdFromNode(file, node, cmdList))
+        })
+        return cmds
+    }
+
+    private getCmdFromNode(file: string, node: latexParser.Node, cmdList: string[] = []): Suggestion[] {
+        const cmds: Suggestion[] = []
+        if (latexParser.isDefCommand(node)) {
+           const name = node.token.slice(1)
+            if (!cmdList.includes(name)) {
+                const cmd: Suggestion = {
+                    label: `\\${name}`,
+                    kind: vscode.CompletionItemKind.Function,
+                    documentation: '`' + name + '`',
+                    insertText: new vscode.SnippetString(name + this.getArgsFromNode(node)),
+                    filterText: name,
+                    package: ''
+                }
+                cmds.push(cmd)
+                cmdList.push(name)
+            }
+        } else if (latexParser.isCommand(node)) {
+            if (!cmdList.includes(node.name)) {
+                const cmd: Suggestion = {
+                    label: `\\${node.name}`,
+                    kind: vscode.CompletionItemKind.Function,
+                    documentation: '`' + node.name + '`',
+                    insertText: new vscode.SnippetString(node.name + this.getArgsFromNode(node)),
+                    filterText: node.name,
+                    package: ''
+                }
+                if (isTriggerSuggestNeeded(node.name)) {
+                    cmd.command = { title: 'Post-Action', command: 'editor.action.triggerSuggest' }
+                }
+                cmds.push(cmd)
+                cmdList.push(node.name)
+            }
+            if (['newcommand', 'renewcommand', 'providecommand', 'DeclarePairedDelimiter', 'DeclarePairedDelimiterX', 'DeclarePairedDelimiterXPP'].includes(node.name) &&
+                Array.isArray(node.args) && node.args.length > 0) {
+                const label = (node.args[0].content[0] as latexParser.Command).name
+                let args = ''
+                if (latexParser.isOptionalArg(node.args[1])) {
+                    const numArgs = parseInt((node.args[1].content[0] as latexParser.TextString).content)
+                    for (let i = 1; i <= numArgs; ++i) {
+                        args += '{${' + i + '}}'
+                    }
+                }
+                if (!cmdList.includes(label)) {
+                    const cmd: Suggestion = {
+                        label: `\\${label}`,
+                        kind: vscode.CompletionItemKind.Function,
+                        documentation: '`' + label + '`',
+                        insertText: new vscode.SnippetString(label + args),
+                        filterText: label,
+                        package: 'user-defined'
+                    }
+                    cmds.push(cmd)
+                    this.definedCmds[label] = {
+                        file,
+                        location: new vscode.Location(
+                            vscode.Uri.file(file),
+                            new vscode.Position(node.location.start.line - 1, node.location.start.column))
+                    }
+                    cmdList.push(label)
+                }
+            }
+        }
+        if (latexParser.hasContentArray(node)) {
+            return cmds.concat(this.getCmdFromNodeArray(file, node.content, cmdList))
+        }
+        return cmds
+    }
+
+    private getArgsFromNode(node: latexParser.Node): string {
+        let args = ''
+        if (!('args' in node)) {
+            return args
+        }
+        let index = 0
+        if (latexParser.isCommand(node)) {
+            node.args.forEach(arg => {
+                ++index
+                if (latexParser.isOptionalArg(arg)) {
+                    args += '[${' + index + '}]'
+                } else {
+                    args += '{${' + index + '}}'
+                }
+            })
+            return args
+        }
+        if (latexParser.isDefCommand(node)) {
+            node.args.forEach(arg => {
+                ++index
+                if (latexParser.isCommandParameter(arg)) {
+                    args += '{${' + index + '}}'
+                }
+            })
+            return args
+        }
+        return args
+    }
+
+    getCmdFromContent(file: string, content: string): Suggestion[] {
+        const cmdReg = /\\([a-zA-Z@_]+(?::[a-zA-Z]*)?)({[^{}]*})?({[^{}]*})?({[^{}]*})?/g
+        const cmds: Suggestion[] = []
+        const cmdList: string[] = []
+        let explSyntaxOn: boolean = false
+        while (true) {
+            const result = cmdReg.exec(content)
+            if (result === null) {
+                break
+            }
+            if (result[1] === 'ExplSyntaxOn') {
+                explSyntaxOn = true
+                continue
+            } else if (result[1] === 'ExplSyntaxOff') {
+                explSyntaxOn = false
+                continue
+            }
+
+
+            if (!explSyntaxOn) {
+                const len = result[1].search(/[_:]/)
+                if (len > -1) {
+                    result[1] = result[1].slice(0, len)
+                }
+            }
+            if (cmdList.includes(result[1])) {
+                continue
+            }
+            const cmd: Suggestion = {
+                label: `\\${result[1]}`,
+                kind: vscode.CompletionItemKind.Function,
+                documentation: '`' + result[1] + '`',
+                insertText: new vscode.SnippetString(this.getArgsFromRegResult(result)),
+                filterText: result[1],
+                package: ''
+            }
+            if (isTriggerSuggestNeeded(result[1])) {
+                cmd.command = { title: 'Post-Action', command: 'editor.action.triggerSuggest' }
+            }
+            cmds.push(cmd)
+            cmdList.push(result[1])
+        }
+
+        const newCommandReg = /\\(?:(?:(?:re|provide)?(?:new)?command)|(?:DeclarePairedDelimiter(?:X|XPP)?))(?:{)?\\(\w+)/g
+        while (true) {
+            const result = newCommandReg.exec(content)
+            if (result === null) {
+                break
+            }
+            if (cmdList.includes(result[1])) {
+                continue
+            }
+
+            const cmd: Suggestion = {
+                label: `\\${result[1]}`,
+                kind: vscode.CompletionItemKind.Function,
+                documentation: '`' + result[1] + '`',
+                insertText: result[1],
+                filterText: result[1],
+                package: 'user-defined'
+            }
+            cmds.push(cmd)
+            cmdList.push(result[1])
+
+            this.definedCmds[result[1]] = {
+                file,
+                location: new vscode.Location(
+                    vscode.Uri.file(file),
+                    new vscode.Position(content.substr(0, result.index).split('\n').length - 1, 0))
+            }
+        }
+
+        return cmds
+    }
+
+    private getArgsFromRegResult(result: RegExpExecArray): string {
+        let text = result[1]
+
+        if (result[2]) {
+            text += '{${1}}'
+        }
+        if (result[3]) {
+            text += '{${2}}'
+        }
+        if (result[4]) {
+            text += '{${3}}'
+        }
+        return text
+    }
+
+}
